version5

+++ src/main/java/com/meli/galaxias/server/core/job/ICalculable.java
@@ -8,6 +8,6 @@ import com.meli.galaxias.server.core.model.Sun;
 public interface ICalculable {
 	public Sun getSun();
 	public List<Planet> getPlanets();
-	public void setSimlateDay(long numberDay);
+	public void setSimlateDay(double numberDay);
 	public String getCode();
 }
