+++ src/main/java/com/meli/galaxias/server/core/model/Planet.java
@@ -24,7 +24,7 @@ public abstract class Planet extends CelestialBody{
 		this.radio = radio;
 	}
 
-	public void setDay(long numberDay) {
+	public void setDay(double numberDay) {
 		System.out.println("Calculando posicion para el dia: " + String.valueOf(numberDay));
 		double gredees = gredeesPerDay * numberDay;
 		double x = Math.sin(gredees) * radio;
